[refactor] The old coq-build is dead!
What was needed in order to run unit tests in Node.js is
now located in `headless.ts`.
It is invoked via the regular CLI (bundled with Webpack).

Fixes #195.
