+++ package.json
@@ -1,15 +1,17 @@
 {
   "name": "jscoq",
-  "version": "0.12.1",
+  "version": "0.12.2",
   "description": "A port of Coq to JavaScript -- run Coq in your browser",
   "bin": {
     "jscoq": "./cli.js"
   },
   "dependencies": {
     "array-equal": "^1.0.0",
+    "assert": "^2.0.0",
     "bootstrap": "^3.4.1",
     "codemirror": "^5.49.2",
     "commander": "^5.0.0",
+    "fflate-unzip": "^0.3.6-0",
     "find": "^0.3.0",
     "glob": "^7.1.3",
     "jquery": "^3.5.1",
@@ -23,6 +25,8 @@
     "vue-context": "^5.1.0"
   },
   "devDependencies": {
+    "@types/find": "^0.2.1",
+    "@types/mkdirp": "^1.0.1",
     "@types/node": "^13.11.1",
     "css-loader": "^3.4.2",
     "file-loader": "^6.0.0",
@@ -34,7 +38,7 @@
     "typescript": "^3.8.3",
     "vue-loader": "^15.9.3",
     "vue-template-compiler": "^2.6.12",
-    "webpack": "^4.44.2",
+    "webpack": "^5.19.0",
     "webpack-cli": "^3.3.11"
   },
   "repository": {
